backend/requirements.txt
# =============================================================================
# Jeseci Smart Learning Academy - Backend Dependencies
# =============================================================================
# Core dependencies for the learning platform backend
# =============================================================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database Drivers
psycopg2-binary>=2.9.11
neo4j>=5.14.0

# Configuration
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP Client (for external APIs)
httpx>=0.26.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
